+++ rust/crd/src/lib.rs
@@ -33,6 +33,7 @@ use strum::{Display, EnumIter, EnumString};
 
 pub const DOCKER_IMAGE_BASE_NAME: &str = "kafka";
 pub const APP_NAME: &str = "kafka";
+pub const OPERATOR_NAME: &str = "kafka.stackable.tech";
 // ports
 pub const CLIENT_PORT_NAME: &str = "kafka";
 pub const CLIENT_PORT: u16 = 9092;