+++ rust/operator/src/utils.rs
@@ -1,5 +1,9 @@
 use snafu::Snafu;
-use stackable_operator::kube::{core::DynamicObject, runtime::reflector::ObjectRef, Resource};
+use stackable_kafka_crd::{KafkaCluster, APP_NAME, OPERATOR_NAME};
+use stackable_operator::{
+    kube::{core::DynamicObject, runtime::reflector::ObjectRef, Resource},
+    labels::ObjectLabels,
+};
 
 #[derive(Debug, Snafu)]
 #[snafu(display(
@@ -32,5 +36,24 @@ impl<K: Resource<DynamicType = ()>> ObjectRefExt for ObjectRef<K> {
             }
             .fail()
         }
+    }
+}
+
+/// Build recommended values for labels
+pub fn build_recommended_labels<'a>(
+    owner: &'a KafkaCluster,
+    controller_name: &'a str,
+    app_version: &'a str,
+    role: &'a str,
+    role_group: &'a str,
+) -> ObjectLabels<'a, KafkaCluster> {
+    ObjectLabels {
+        owner,
+        app_name: APP_NAME,
+        app_version,
+        operator_name: OPERATOR_NAME,
+        controller_name,
+        role,
+        role_group,
     }
 }